+++ src/components/burger-total/burger-total.js
@@ -1,11 +1,18 @@
-import { useContext } from 'react';
+import { useSelector } from 'react-redux';
+
 
 import { CurrencyIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 
-import { BurgerTotalContext } from '../../utils/burger-api';
+import { burgerConstructorIngredients } from '../app/app';
 
 function BurgerTotal() {
-  const sum = useContext(BurgerTotalContext);
+  const sum = useSelector(burgerConstructorIngredients).constructorList
+              .reduce((acc, curr) => {
+                const price = curr.ingredient.type === 'bun'
+                ? curr.ingredient.price * 2 
+                : curr.ingredient.price;
+                return acc + price;
+              }, 0 );
 
   return (
     <div>
