+++ src/components/ingredient-details/ingredient-details.js
@@ -3,8 +3,9 @@ import { useSelector } from 'react-redux';
 import IngredientDetailsStyles from './ingredient-details.module.css';
 
 function IngredientDetails() {
-  
-  const ingredientData = useSelector(store => store.burgerIngredients.ingredientSelected);
+
+  const burgerIngredientsDetails = store => store.burgerIngredientsDetails;
+  const ingredientData = useSelector(burgerIngredientsDetails).ingredientSelected;
 
   return (
     <div className={IngredientDetailsStyles.content}>
