+++ src/services/actions/burger-constructor-ingredients.js
@@ -1,0 +1,4 @@
+export const CONSTRUCTOR_ADD_INGREDIENT =     'CONSTRUCTOR_ADD_INGREDIENT';
+export const CONSTRUCTOR_REMOVE_INGREDIENT =  'CONSTRUCTOR_REMOVE_INGREDIENT';
+export const CONSTRUCTOR_CLEAR_INGREDIENTS =  'CONSTRUCTOR_CLEAR_INGREDIENTS';
+export const CONSTRUCTOR_SWAP_INGREDIENTS =   'CONSTRUCTOR_SWAP_INGREDIENTS';
