--- src/services/actions/burger-constructor.js
@@ -1,7 +1,0 @@
-export const CONSTRUCTOR_GET_ORDERNUM =         'CONSTRUCTOR_GET_ORDERNUM';
-export const CONSTRUCTOR_GET_ORDERNUM_SUCCESS = 'CONSTRUCTOR_GET_ORDERNUM_SUCCESS';
-export const CONSTRUCTOR_GET_ORDERNUM_FAILED =  'CONSTRUCTOR_GET_ORDERNUM_FAILED';
-export const CONSTRUCTOR_ADD_INGREDIENT =       'CONSTRUCTOR_ADD_INGREDIENT';
-export const CONSTRUCTOR_REMOVE_INGREDIENT =    'CONSTRUCTOR_REMOVE_INGREDIENT';
-export const CONSTRUCTOR_CLEAR_INGREDIENTS =    'CONSTRUCTOR_CLEAR_INGREDIENTS';
-export const CONSTRUCTOR_SWAP_INGREDIENTS =     'CONSTRUCTOR_SWAP_INGREDIENTS';
