+++ src/services/actions/burger-ingredients-details.js
@@ -1,0 +1,2 @@
+export const INGREDIENTS_SELECT_INGREDIENT =    'INGREDIENTS_SELECT_INGREDIENT';
+export const INGREDIENTS_DESELECT_INGREDIENT =  'INGREDIENTS_DESELECT_INGREDIENT';
