--- src/services/actions/burger-ingredients.js
@@ -1,6 +1,0 @@
-export const INGREDIENTS_GET_INGREDIENTS =          'INGREDIENTS_GET_INGREDIENTS';
-export const INGREDIENTS_GET_INGREDIENTS_FAILED =   'INGREDIENTS_GET_INGREDIENTS_FAILED';
-export const INGREDIENTS_GET_INGREDIENTS_SUCCESS =  'INGREDIENTS_GET_INGREDIENTS_SUCCESS';
-
-export const INGREDIENTS_SELECT_INGREDIENT =        'INGREDIENTS_SELECT_INGREDIENT';
-export const INGREDIENTS_DESELECT_INGREDIENT =      'INGREDIENTS_DESELECT_INGREDIENT';
